@@ -18,7 +18,7 @@
 		"classmap": ["tests/"]
 	},
 	"require": {
-		"php": "^7.2 || ~8.0",
+		"php": "~7.4 || ~8.0",
 		"consistence/consistence": "~2.0",
 		"phpstan/phpstan": "^0.12"
 	},
